Big One
A little bit of everything. Biggest part is the inventory system. While it isn't complete, most of its parts are there. I need to start doing this weekly. Scripts aren't formatted for readability, as I've been working on them in stolen moments, but I'll remedy that in my very next refactor, which should be after my first sprint.